+++ CHANGELOG.md
@@ -7,6 +7,13 @@ writing, this is a rolling-release project without any meaningful versioning
 whatsoever. Tags/releases may be created for the sole purpose of documenting
 major updates to the project.
 
+## 2022-12-14
+
+### Fixed
+
+- Clear session if state has expired
+  ([#823](https://github.com/giscus/giscus/pull/823)).
+
 ## 2022-12-10
 
 ### Changed